+++ src/main/controllers/features/sole-trader/correspondenceAddressAutoLookupController.ts
@@ -13,6 +13,7 @@ import { ACSPData } from "../../../model/ACSPData";
 import { Session } from "@companieshouse/node-session-handler";
 import { USER_DATA } from "../../../common/__utils/constants";
 import { saveDataInSession } from "../../../common/__utils/sessionHelper";
+import CorrespondenceAddressAutoLookService from "../../../services/correspondence-address/address-autolookup";
 
 export const get = async (req: Request, res: Response, next: NextFunction) => {
     const session: Session = req.session as any as Session;
@@ -60,60 +61,10 @@ export const post = async (req: Request, res: Response, next: NextFunction) => {
             const correspondencePremise = req.body.premise;
 
             if (correspondencePremise !== "" && ukAddresses.find((address) => address.premise === correspondencePremise)) {
-                let address = {
-                    premise: "",
-                    addressLine1: "",
-                    addressLine2: "",
-                    postTown: "",
-                    postalCode: "",
-                    country: ""
-                };
-                for (const ukAddress of ukAddresses) {
-                    if (ukAddress.premise === correspondencePremise) {
-                        address = {
-                            premise: ukAddress.premise,
-                            addressLine1: ukAddress.addressLine1,
-                            addressLine2: ukAddress.addressLine2!,
-                            postTown: ukAddress.postTown,
-                            postalCode: ukAddress.postcode,
-                            country: getCountryFromKey(ukAddress.country)
-                        };
-                    }
-                }
-                // Save the correspondence address to session
-                const correspondenceAddress : Address = {
-                    propertyDetails: address.premise,
-                    line1: address.addressLine1,
-                    line2: address.addressLine2,
-                    town: address.postTown,
-                    country: address.country,
-                    postcode: address.postalCode
-                };
-                const userAddresses : Array<Address> = ACSPData?.addresses ? ACSPData.addresses : [];
-                userAddresses.push(correspondenceAddress);
-                ACSPData.addresses = userAddresses;
-                saveDataInSession(req, USER_DATA, ACSPData);
-                res.redirect(BASE_URL + SOLE_TRADER_CORRESPONDENCE_ADDRESS_CONFIRM);
-
+                CorrespondenceAddressAutoLookService.saveCorrespondenceAddressToSession(session, req, ukAddresses, correspondencePremise);
+                res.redirect(addLangToUrl(BASE_URL + SOLE_TRADER_CORRESPONDENCE_ADDRESS_CONFIRM, lang));
             } else {
-
-                const addressList : Array<Address> = [];
-                for (const ukAddress of ukAddresses) {
-                    const address = {
-                        propertyDetails: ukAddress.premise,
-                        line1: ukAddress.addressLine1,
-                        line2: ukAddress.addressLine2,
-                        town: ukAddress.postTown,
-                        country: getCountryFromKey(ukAddress.country),
-                        postcode: ukAddress.postcode,
-                        formattedAddress: ukAddress.premise + ", " + ukAddress.addressLine1 + ", " + ukAddress.postTown + ", " + getCountryFromKey(ukAddress.country) + ", " + ukAddress.postcode
-                    };
-
-                    addressList.push(address);
-
-                }
-                ACSPData.addresses = addressList;
-                saveDataInSession(req, USER_DATA, ACSPData);
+                CorrespondenceAddressAutoLookService.saveAddressListToSession(session, req, ukAddresses);
                 const nextPageUrl = addLangToUrl(BASE_URL + SOLE_TRADER_AUTO_LOOKUP_ADDRESS_LIST, lang);
                 res.redirect(nextPageUrl);
